Initial commit of BSIR Admin Tool
The BSIR Admin Tool allows CRUDing a mapping between Users, Applications and Datasets. It publishes that mapping in both an Sqlite3 database as well as a YAML document. Callbacks are installed on the create and destroy handlers for Mapping, and will overwrite the YAML storage with the new state. The YAML storage is never read by the Admin app.

On the list view mappings that have become invalid (missing directories or bad permissions) since creation are called out with a FontAwesome warning icon.

Most of the pseudofun specific code has been removed, but there are bound to be some remanents. The new mapping form is currently a straight post request, so it is not sticky in case of failure. Likewise it would be good to add more information to the valid field in the list view (in either a modal or a tool tip), so that the admin knows exactly what remedy is required.