ensure all files created are user & group writable
This is just a safety measure. Now if I touch a file
in the Rails console that file is user and group writable.

note that this still has no affect on the database created when
running rake db:migrate, even though that task depends on the
environment so care must be taken to chmod 775 that file after it is created